Refactor: remove BotAPI interface from bot_api.go and consolidate its definition in service.go
