--- pkg/bot/bot_api.go
@@ -1,11 +1,0 @@
-package bot
-
-import tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
-
-// BotAPI interface represents the Telegram bot API capabilities we use
-type BotAPI interface {
-	Send(c tgbotapi.Chattable) (tgbotapi.Message, error)
-	Request(c tgbotapi.Chattable) (*tgbotapi.APIResponse, error)
-	StopReceivingUpdates()
-	GetUpdatesChan(config tgbotapi.UpdateConfig) tgbotapi.UpdatesChannel
-}
